test: move corpus extraction out of shared conftest into corpus module
The corpus extraction (4.1 MB tarball, 7,462 files) and per-request
param building lived at module level in tests/integration/conftest.py.
A conftest loads for every invocation that touches its directory, so
`make integration-tests` — which runs only test_cedar_integration_tests.py
— paid the full extraction + json.load cost at collection time:
collection of the 74 hand-written tests took ~10.4s wall (6.1s sys).

Move CORPUS_TARBALL / _extract_corpus_once / CORPUS_DIR /
_list_corpus_test_params / CORPUS_PARAMS and the corpus_dir fixture into
test_cedar_corpus_tests.py. pytest only imports a test module when it is
actually being collected, so `make integration-tests` no longer touches
any corpus code while `make corpus-tests` is unaffected. conftest.py held
nothing else and is removed.

Collection of the hand-written suite drops from ~10.4s to ~0.55s.
